model/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.6.0
asttokens==2.4.1
astunparse==1.6.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
contourpy==1.2.1
crypto==1.4.1
cryptography==39.0.0
cycler==0.12.1
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.8
Django==4.1.5
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
firebase==3.0.1
Flask==2.1.2
flatbuffers==24.3.25
fonttools==4.53.1
gast==0.6.0
gcloud==0.18.3
google-pasta==0.2.0
googleapis-common-protos==1.57.1
greenlet==3.0.3
grpcio==1.66.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.21.0
httptools==0.6.1
httpx==0.27.0
idna==3.4
image==1.5.33
ipython==8.27.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
jwcrypto==1.4.2
keras==3.5.0
kiwisolver==1.4.5
libclang==18.1.1
lightgbm==4.5.0
mariadb==1.1.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.0
mysql-connector-python==9.0.0
Naked==0.1.32
namex==0.0.8
numpy==1.24.1
oauth2client==4.1.3
opencv-python==4.7.0.68
openpyxl==3.1.1
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.3
packaging==24.1
pandas==1.5.3
parso==0.8.4
Pillow==9.4.0
pipenv==2024.0.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.21.12
pure_eval==0.2.3
py-crypto==0.0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.0
pydantic_core==2.18.4
Pygments==2.18.0
PyMySQL==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jwt==4.0.0
python-multipart==0.0.9
pytz==2022.7.1
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.10.1
rich==13.7.1
rsa==4.9
scikit-learn==1.5.1
scikit-surprise==1.1.4
scipy==1.14.0
seaborn==0.13.2
shellescape==3.8.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.4.3
stack-data==0.6.3
starlette==0.37.2
surprise==0.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.1
tzdata==2022.7
ujson==5.10.0
urllib3==1.26.13
uvicorn==0.30.1
virtualenv==20.26.3
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==2.1.2
wrapt==1.14.1
xgboost==2.1.1
xlrd==2.0.1
